App version: 0.2
TaskList now uses a LinkedList<String> for commandOutput, thus allowing to iterate and look for specific entries.
Fixed a bug on Linux platform which caused printProcess() to print true for some process even when not running.
App arguments can now be accessed from all methods in Main class.
Added support for --menu [int] (for pre-choosing menu entries, can be queued for multiple selection), --pair-device [string] (for pre-choosing a pair device. It must be called along with --menu 0 for a complete automation)